forgot to move to double for centerOnRegion
git-svn-id: file:///home/hoijui/src/IsaViz/tmp2/zvtm.svn/code@3585 74e165f3-c697-40a6-91af-552c8936d7fe

+++ zvtm/trunk/src/main/java/fr/inria/zvtm/engine/portals/CameraPortal.java
@@ -260,7 +260,7 @@ public class CameraPortal extends Portal {
 		return new Location(dx, dy, newAlt);
     }
         
-    public Location centerOnRegion(int d, long x1, long y1, long x2, long y2){
+    public Location centerOnRegion(int d, double x1, double y1, double x2, double y2){
         Location l = centerOnRegion(x1, y1, x2, y2);
 		Animation trans = 
 		    VirtualSpaceManager.INSTANCE.getAnimationManager().getAnimationFactory().
